ポート転送のファイルを一つにした
ついでにエラーのバリアント名を調整
